LPS-51335 Move
